backend/requirements.txt
fastapi
uvicorn[standard]

numpy==1.26.4
pandas==2.3.0
pyyaml==6.0.2

fastparquet==2024.11.0
pyarrow==20.0.0

# PyTorch estável para Python 3.11 + macOS (CPU)
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

tqdm
python-dotenv

# LangChain eco
langchain>=0.3.26
langchain-community==0.3.27
langchain-core
langchain-chroma==0.2.4
langchain-openai==0.3.27

# Sentence Transformers (usa HuggingFace Transformers)
sentence-transformers==5.0.0

# Integração com PostgreSQL
sqlalchemy>=2.0.0
psycopg2-binary

# Consultas analíticas locais
duckdb>=1.1.0
